add different testfiles, add function to check environmentvariables
