+++ src/models/environmentVariable.js
@@ -1,0 +1,30 @@
+/*
+ * Copyright WebGate Consulting AG, 2020
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ *
+ */
+export class EnvironmentVariable {
+  constructor() {
+    this.variablename = null;
+    this.value = null;
+    this.descr = null;
+  }
+  static createEnvironmentVariable(variablename) {
+    const variable = new EnvironmentVariable();
+    variable.variablename = variablename;
+    variable.value = null;
+    variable.descr = null;
+    return variable;
+  }
+}
